refactor envelope with clap derive api
