+++ src/ui/table/mod.rs
@@ -1,5 +1,4 @@
 pub mod arg;
-pub mod args;
 pub mod table;
 
 pub use table::*;
